requirements.txt
discord.py
python-dotenv
requests
aiohttp
beautifulsoup4
lxml